Fixed sticky table headers on Safari 😒
NOTE: There is a slight trade-off for this — the table headers will show their border-radiuses when scrolling. This is obviously still an upgrade for Safari users, but a downgrade for other browsers. However, it's probably worth it, since Safari is the most popular browser used to access the site (just above Chrome), and having it completely broken for ~33% of visitors is better than having slightly weird looking borders but working for 100% of users.