Show times in Australia/Sydney timezone rather than local timezone
Rationale: The data from NSW Health is given in the Australia/Sydney timezone and therefore all dates in the chart etc. are assumed to be Australia/Sydney days, it would be weird if all the date windows were shifted around based on the user's timezone leading to different daily counts. Therefore, it makes sense to have the last updated date shown in Australia/Sydney as well, so it's always consistent with temporalCoverageTo regardless of the timezone.

This also leads to a 'fix' for some actual numbers as well as the actual last updated time/date displayed — in timezones that are so far behind that the last updated timestamp is the previous day in that timezone, everything was previously calculated for essentially the day before the data was updated. Eg. if the data was updated on Friday (so valid till Thursday), the last updated date would be shown as Thursday and the valid date would be shown as Wednesday, and in addition, the numbers would all be calculated assuming Wednesday was "today", so all with incorrect today/this week numbers. Obviously this 'bug' doesn't affect the large majority of visitors since most would be in Australia or another Australian timezone which would still have the same date most of the time, but it does fix it for users in places like Hawaii.