Fix selecting PostcodePicker options in Safari
The bug was because Safari doesn't give focus to buttons when pressed, therefore event.relatedTarget is undefined.
https://bugs.webkit.org/show_bug.cgi?id=22261
This bug was reported back in 2008(!!!) 😠 #SafariIsTheNewIE

The mousedown workaround is a bit of an ugly hack because now navigation will happen on mousedown rather than mouseup/click, but until I can find a better solution, it's better than how it is currently.